% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gamma_params}
\alias{gamma_params}
\title{Estimate the parameters for a Gamma distribution}
\usage{
gamma_params(x)
}
\arguments{
\item{x}{a data vector to operate on}
}
\value{
a list of the shape (alpha) and rate (beta) parameters
  and the mean and variance
}
\description{
This is a simple function to estimate what the parameters for a Gamma
distribution would be from a data vector.  It is used internally to
generate start values.
}
\author{
Joshua F. Wiley <josh@elkhartgroup.com>
}
\keyword{utilities}

