% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{pval_smartformat}
\alias{pval_smartformat}
\title{nice formatting for p-values}
\usage{
pval_smartformat(p, d = 3, sd = 5)
}
\arguments{
\item{p}{a numeric pvalue}

\item{d}{the digits less than which should be displayed as less than}

\item{sd}{scientific digits for round}
}
\description{
nice formatting for p-values
}
\examples{
varian:::pval_smartformat(c(1, .15346, .085463, .05673, .04837, .015353462,
  .0089, .00164, .0006589, .0000000053326), 3, 5)
}
\author{
Joshua F. Wiley <josh@elkhartgroup.com>
}
\keyword{utilities}

