% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summary.vm}
\alias{summary.vm}
\title{Summary method for variability model objects}
\usage{
\method{summary}{vm}(object, digits = getOption("digits"), ...)
}
\arguments{
\item{object}{An object of class \code{vm}.}

\item{digits}{The number of digits to use for rounding.}

\item{\dots}{Additional arguments. Not currently used.}
}
\description{
Summary method for variability model objects
}

