% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{VM}
\alias{VM}
\alias{VM-package}
\title{Variability Models in R}
\description{
The VM package estimates the variability in a repeated
  measure outcome and uses that as an outcome or a predictor in a second stage
  model.
}
\details{
The VM package has one primary function:
\enumerate{
  \item \code{\link{vm_predict}}
}

\tabular{ll}{
Package: \tab VM\cr
Type: \tab Package\cr
Version: \tab 0.6-2\cr
Date: \tab 2013-10-24\cr
License: \tab LGPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Joshua F. Wiley \email{josh@elkhartgroup.com}

Maintainer: Joshua F. Wiley \email{josh@elkhartgroup.com}
}
\references{
Elkhart Group Ltd. \url{http://elkhartgroup.com}
}
\seealso{
See \code{\link{vm_predict}} for an example running a model.
}
\keyword{package}

