% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vm_stan}
\alias{vm_stan}
\title{Create a Stan class VM model}
\usage{
vm_stan(type = c("vm_predict", "vm_predict_mediation"), useU = TRUE, ...)
}
\arguments{
\item{type}{A character string indicating the type of model,
currently only \dQuote{vm_predict} and \dQuote{vm_predict_mediation}.}

\item{useU}{A logical value whether the latent intercept estimated in Stage 1 should
also be used as a predictor.  Defaults to \code{TRUE}.}

\item{\dots}{Additional arguments passed to \code{stan_model}.}
}
\value{
A compiled Stan model.
}
\description{
Internal function to create and compile a Stan model.
}
\examples{
# Make Me!
\dontrun{
  test1 <- vm_stan("vm_predict", useU=TRUE)
  test2 <- vm_stan("vm_predict", useU=FALSE)
  test3 <- vm_stan("vm_predict_mediation", useU=TRUE)
  test4 <- vm_stan("vm_predict_mediation", useU=FALSE)
}
}
\author{
Joshua F. Wiley <josh@elkhartgroup.com>
}
\keyword{models}

