% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeler.R
\name{vm_stan}
\alias{vm_stan}
\title{Create a Stan class VM object}
\usage{
vm_stan(design = c("V -> Y", "V -> M -> Y", "V", "X -> V", "X -> V -> Y",
  "X -> M -> V"), useU = TRUE, UQ = FALSE, IIVQ = FALSE, centerU = 0,
  centerIIV = 0, ..., template_only = FALSE)
}
\arguments{
\item{design}{A character string indicating the type of model to be run.  One of
\dQuote{V -> Y} for variability predicting an outcome,
\dQuote{V -> M -> Y} for mediation of variability on an outcome,
\dQuote{V} to take posterior samples of individual variability estimates alone.}

\item{useU}{A logical value whether the latent intercept estimated in Stage 1 should
also be used as a predictor.  Defaults to \code{TRUE}.  Note if there is a
mediator as well as main outcome, the latent intercepts will be used as a predictor
for both.}

\item{UQ}{A logical value whether the latent intercept estimated in Stage 1 should
also be used as a predictor with a qudratice effect.  Defaults to \code{FALSE}.
Note if there is a mediator as well as main outcome, the latent intercepts
will be used as a predictor for both.}

\item{IIVQ}{A logical value whether the latent variabilities estimated in Stage 1 should
also be used as a predictor with a qudratice effect.  Defaults to \code{FALSE}.
Note if there is a mediator as well as main outcome, the latent intercepts
will be used as a predictor for both.}

\item{centerU}{A numeric vector of length one (scalar) that is used to center the
latent intercept estimates before using as a predictor of the outcome / mediator.
Uses the formula: (U - centerU). Particularly useful when including quadratic terms.}

\item{centerIIV}{A numeric vector of length one (scalar) that is used to center the
latent variability estimates before using as a predictor of the outcome / mediator.
Uses the formula: (IIV - centerIIV). Particularly useful when including quadratic terms.}

\item{template_only}{A logical, not commonly used. If \code{TRUE} returns
the Stan template code only, without actually compiling the Stan model.
Useful for modifying the default models.}

\item{\dots}{Additional arguments passed to \code{stan_model}.}
}
\value{
A compiled Stan model.
}
\description{
Internal function to create and compile a Stan model.
}
\examples{
varian:::vm_stan("V -> Y", useU=TRUE, template_only = TRUE)
varian:::vm_stan("V -> Y", useU=TRUE, UQ = TRUE, IIVQ = TRUE, template_only = TRUE)
varian:::vm_stan("V -> M -> Y", useU=TRUE, UQ = TRUE, IIVQ = TRUE, template_only = TRUE)
}
\author{
Joshua F. Wiley <josh@elkhartgroup.com>
}
\keyword{models}

